+++ ReleaseNotes.md
@@ -2,7 +2,7 @@
 
 
 ## Version 2.6.2
-- recursive blocking implemented for solving contunuous Sylvester equations in `sylvc` for `BlasFloat` data
+- recursive blocking implemented for solving continuous and discrete Sylvester equations in `sylvc` and `sylvd` for `BlasFloat` data
 
 ## Version 2.6.1
 Several enhancements of the generalized discrete Lyapunov equation solver `dlyap` and of the